third_party/blink/web_tests/fast/forms/time-multiple-fields/time-multiple-fields-preserve-value-after-history-back-expected.txt

Checks input value is preserved after backward and forward

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Went back to a page. Checking a form control which had a full value:
PASS "01:01" is "01:01"
Went back to a page. Checking a form control which had a partial value:
FAIL "01:--:-- AM" should be 01:--:-- AM. Was 01:--:-- AM.
Went forward to a page. Checking a form control which had a full value:
PASS "02:02" is "02:02"
Went forward to a page. Checking a form control which had a partial value:
FAIL "--:59:-- --" should be --:59:-- --. Was --:59:-- --.


PASS successfullyParsed is true

TEST COMPLETE
